+++ contracts/fork-test.sh
@@ -74,7 +74,7 @@ main()
         if [[ $FORK_NETWORK_NAME == "holesky" ]]; then
             # Run all files with `.holesky.fork-test.js` suffix when no file name param is given
             # pass all other params along
-            params+="test/**/*.holesky.fork-test.js"
+            params+="test/**/*.holesky-fork-test.js"
         else
             # Run all files with `.fork-test.js` suffix when no file name param is given
             # pass all other params along
