--- contracts/test/strategies/nativeSsvStaking.holesky.fork-test.js
@@ -1,147 +1,0 @@
-const { expect } = require("chai");
-const { utils, BigNumber } = require("ethers");
-const addresses = require("../../utils/addresses");
-const { units, oethUnits } = require("../helpers");
-
-const { loadSimpleOETHFixture } = require("./../_fixture");
-
-describe("Holesky ForkTest: Native SSV Staking Strategy", function () {
-  this.timeout(0);
-
-  // Retry up to 3 times on CI
-  // this.retries(isCI ? 3 : 0);
-
-  let fixture;
-  beforeEach(async () => {
-    fixture = await loadSimpleOETHFixture();
-  });
-
-  describe("Initial setup", function () {
-    it("Should verify the initial state", async () => {
-      const { nativeStakingSSVStrategy } = fixture;
-      await expect(
-        await nativeStakingSSVStrategy.WETH_TOKEN_ADDRESS()
-      ).to.equal(addresses.holesky.WETH, "Incorrect WETH address set");
-      await expect(await nativeStakingSSVStrategy.SSV_TOKEN_ADDRESS()).to.equal(
-        addresses.holesky.SSV,
-        "Incorrect SSV Token address"
-      );
-      await expect(
-        await nativeStakingSSVStrategy.SSV_NETWORK_ADDRESS()
-      ).to.equal(addresses.holesky.SSVNetwork, "Incorrect SSV Network address");
-    });
-
-    it("Should check that the fuse interval is configured correctly", async () => {
-      const { nativeStakingSSVStrategy } = fixture;
-
-      await expect(utils.parseEther("21.6")).to.equal(
-        await nativeStakingSSVStrategy.fuseIntervalStart()
-      );
-      await expect(utils.parseEther("25.6")).to.equal(
-        await nativeStakingSSVStrategy.fuseIntervalEnd()
-      );
-    });
-  });
-
-  describe("Deposit/Allocation register validators", function () {
-    it("Should mint using WETH and see balance on the strategy contract", async () => {
-      const { josh, weth } = fixture;
-      await mintTest(fixture, josh, weth, "32");
-    });
-
-    it("Should mint using WETH and deposit to a validator", async () => {
-      const { josh, weth } = fixture;
-      await mintTest(fixture, josh, weth, "32");
-
-      await registerAndDepositTest(fixture);
-    });
-  });
-
-  describe("Withdraw", function () {});
-
-  describe("Balance/Assets", function () {});
-
-  const registerAndDepositTest = async (fixture) => {
-    const { nativeStakingSSVStrategy, strategist } = fixture;
-
-    await nativeStakingSSVStrategy
-      .connect(strategist) // TODO this will be a Defender relayer
-      .registerSsvValidator(
-        // pubkey
-        "0xb279166b18ca9ced3e0a83385799ae33b27c6cd2ad2083b9b7c33c555d091933a09b1af9fb65c6c4d51c40ca931447a9",
-        // operator ids
-        [111, 119, 230, 252],
-        // shares data
-        "0xb1edfddfdf70c8c56cfba3f00ae09815db2383c4faa328733df73dd4616492a58648d543642078c72e2bad60a22322d70ea44cc953c05aa0404681fb57498daa9cbd5c8cc98f736c75a8bf908f0eb98d6448b6778f35b1ce724fee8a1bb53f9ab60f29c5fc7c53e1577a27581b8aff689ac2fdeda72e9620f6991edb7779acef68ed2e6ea4ccfaecd00ccac09b7ed8b7abc72ea5c5469e1b5603c73fe3d054e074c88446c4e31abbc34ee309022524da5aacecb4aa83bf0ac658931813a402e4a4b1d9f83d54404b730e04e61dc82e872fa397e3a9ea249715c861ef0d4b309fdba762cd07f0506925a7a4e3e52d2c0e8f07ae01b8bb393df14c17648340c55e3a82564efb943de4033308d197d562c8bf377a589c87e4c7757afe5964ec92a616622138797c4a647dda550e3b94e3a6152cdda20d4a7b491218ab7df46a2eb9c3963811bf0555b272bf4ef5a8c0c2496e133d1cc34b01c7a5cb40d379e607d0bd7f0d8fcb965ab8d5c7634d80dbd3fac67227b53ec6fbd1a5207bfea8cb202ac88fef732328a2b717b6284af9f34a251e5ecf5eb744a4cf4df35abb423ca02556df957cd8bc645b83d497235b92f310996748c54a89106937cfcf182744ad423b104ca0a61a351d15aa9ae98093f64de31e14a7203304021ebbe2ee0e3f91e1641ff7da84396a83643101cfe67640c05467ffa8de458ebb4db0fcffe9210b72c50f8835cb636e1f5e7356092f15b29fb11761ce36a6601d599eb02d82aa02ce300a7e6c538ca72133036e5f3634c29b27c66964f65086e3f725d535f5e7d92f6a0b589a71caa7d0ca572c5192fe6ef5a4b44ad05cbb8c214a444a399b035263a1d2853bfe43b457873e700be599d3ff03532df1f52be09f6607acfc466f8bb4c4d98d74908c7394d516b738a6143a5456dc43a45f1f00a6515e11b5efc6f4dabfb3d967941ac159a8b6f7c464760c4770082b1d97600cabb1dc286d20c321e6bcb36138507915c399d8639d8a66143bff1f5a6d6cb63d0a8b560134dab42e9d23100c05e51104e068c0c76ecb5382d04e34375efea6467dfb096f6dc12b67609ea1d8310044ab8e80f75341b4aa3eb8dd5f465a3dc6b3f1dea8f25c77cdc34f0eb04247c1ac73bde467c8048cc4d57afefcb33de7520048baeaa9401fc175061881d5c60c89a4fe71cf6af8ed7ba3f6b5da42ef2ee5454f422ecf0b02d09f1ba77d35b56240232237810ffe4ff8e49c4a40a9aab7880ea91472440f30b22797f97d90a68e7904513a4267af47d988d96a17c5a2e90c4ad7c6fb56de8ba7e5488cfc16d1015963e2778e2b3def5ffdda1439cf8750b5823e31f4ba59f1eaf17c7ee98a69495a3c38c177175554b343c21f91a708c6313f146bbdde18b1fcead4e0d0274c8c1010c96f79b81060b850ab5c1da001fc908e45b84d57837fbbd272a4349261e500ce56560939cba45a58f2d0bdba3c6631e52d4e0b7a29ea50f14ed36c1ccb5fca6fdc45f881764b3ee39949ef5df255d9afdedfdf390409dadd31df7540d25174cf21a33dce1a2fd6097967be67267aa7e9353db71821bd89db0b3887eebc01cb64a4116edefac323fdde618a34d91545bab38f920e8e6f38b22a3a243b261fd56f0ec1ff6187b3ed42ddeeadf2c7a78c154d44ab332e293ca2b03e77ae1fe628378e336bc73149df4d7cbc86df73d04b9d35a9fe6a87a865f92337a8df10d5a5cf4dcc6cfba73bdd9b13d3b671acfc829b87f869ed42a48ced74099f92ac79721a79ac93d7c4d5e9be780fe1a78f807fd6a222fac05c3c8b9cd4182cb84617abaa72815b5dceec1d58b1278bf90498e4be3a456428866170046c",
-        // amount
-        BigNumber.from("1534241968000000000"),
-        // cluster tuple
-        [0, 0, 0, true, 0]
-      );
-
-    await nativeStakingSSVStrategy
-      .connect(strategist) // TODO this will be a Defender relayer
-      .stakeEth([
-        [
-          //pubkey
-          "0xb279166b18ca9ced3e0a83385799ae33b27c6cd2ad2083b9b7c33c555d091933a09b1af9fb65c6c4d51c40ca931447a9",
-          //signature
-          "0xa450d596551c7fb7aca201e9a075b034d8da1ec7bf8806740ca53c0e8653465ed9cd26d6ce10290581586676eb0dd896022a243dc42179337c9c4c2a60969a11bb9e4a2dcf57a783daf880999f6db34d1e42163cb96287b3bb91b03361942b80",
-          //depositDataRoot
-          "0x3f327f69bb527386ff4c2f820e6e375fcc632b1b7ee826bd53d4d2807cfd6769",
-        ],
-      ]);
-  };
-
-  const mintTest = async (fixture, user, asset, amount = "32") => {
-    const { oethVault, oeth, weth, nativeStakingSSVStrategy, strategist } =
-      fixture;
-
-    const unitAmount = await units(amount, asset);
-
-    if (asset.address != weth.address) {
-      const tx = oethVault.connect(user).mint(asset.address, unitAmount, "0");
-      await expect(tx).to.be.revertedWith("Unsupported asset for minting");
-      return;
-    }
-
-    await oethVault.connect(user).allocate();
-    await oethVault.connect(user).rebase();
-
-    const currentSupply = await oeth.totalSupply();
-    const currentBalance = await oeth.connect(user).balanceOf(user.address);
-    const currentStrategyBalance = await nativeStakingSSVStrategy.checkBalance(
-      weth.address
-    );
-
-    // Mint OETH w/ asset
-    await oethVault.connect(user).mint(asset.address, unitAmount, 0);
-
-    await oethVault
-      .connect(strategist)
-      .depositToStrategy(
-        nativeStakingSSVStrategy.address,
-        [asset.address],
-        [unitAmount]
-      );
-
-    await oethVault.connect(user).allocate();
-
-    const newBalance = await oeth.connect(user).balanceOf(user.address);
-    const newSupply = await oeth.totalSupply();
-    const newStrategyBalance = await nativeStakingSSVStrategy.checkBalance(
-      weth.address
-    );
-
-    const balanceDiff = newBalance.sub(currentBalance);
-    // Ensure user has correct balance (w/ 1% slippage tolerance)
-    expect(balanceDiff).to.approxEqualTolerance(oethUnits(amount), 2);
-
-    // Supply checks
-    const supplyDiff = newSupply.sub(currentSupply);
-    const oethUnitAmount = oethUnits(amount);
-    expect(supplyDiff).to.approxEqualTolerance(oethUnitAmount, 1);
-
-    expect(unitAmount).to.equal(newStrategyBalance.sub(currentStrategyBalance));
-  };
-});
